+++ code/core/src/components/components/tooltip/ListItem.tsx
@@ -123,6 +123,7 @@ const Item = styled.div<ItemProps>(
   ({ theme }) => ({
     width: '100%',
     border: 'none',
+    borderRadius: theme.appBorderRadius,
     background: 'none',
     fontSize: theme.typography.size.s1,
     transition: 'all 150ms ease-out',
@@ -147,6 +148,15 @@ const Item = styled.div<ItemProps>(
       },
       '&:hover svg': {
         opacity: 1,
+      },
+    },
+  ({ theme, as }) =>
+    as === 'label' && {
+      '&:has(input:not(:disabled))': {
+        cursor: 'pointer',
+        '&:hover': {
+          background: theme.background.hoverable,
+        },
       },
     },
   ({ disabled }) => disabled && { cursor: 'not-allowed' }