+++ code/core/src/manager/components/sidebar/TestingModule.stories.tsx
@@ -1,12 +1,20 @@
 import React from 'react';
 
+import type { Listener } from '@storybook/core/channels';
+import { styled } from '@storybook/core/theming';
 import { Addon_TypesEnum } from '@storybook/core/types';
 import type { Meta, StoryObj } from '@storybook/react';
-import { fn, userEvent } from '@storybook/test';
+import { fireEvent, fn } from '@storybook/test';
 
-import type { TestProviders } from '@storybook/core/core-events';
+import { TESTING_MODULE_CONFIG_CHANGE, type TestProviders } from '@storybook/core/core-events';
+import { ManagerContext, mockChannel } from '@storybook/core/manager-api';
 
 import { TestingModule } from './TestingModule';
+
+const TestProvider = styled.div({
+  padding: 8,
+  fontSize: 12,
+});
 
 const baseState = {
   details: {},
@@ -42,15 +50,33 @@ const testProviders: TestProviders[keyof TestProviders][] = [
     type: Addon_TypesEnum.experimental_TEST_PROVIDER,
     id: 'linting',
     name: 'Linting',
-    title: () => 'Linting',
-    description: () => 'Watching for changes',
+    render: () => <TestProvider>Custom render function</TestProvider>,
     ...baseState,
     watching: true,
   },
 ];
 
+let triggerUpdate: () => void;
+const channel = mockChannel();
+const managerContext: any = {
+  api: {
+    on: (eventName: string, listener: Listener) => {
+      if (eventName === TESTING_MODULE_CONFIG_CHANGE) {
+        triggerUpdate = listener;
+      }
+      return channel.on(eventName, listener);
+    },
+    off: (eventName: string, listener: Listener) => channel.off(eventName, listener),
+    runTestProvider: fn().mockName('api::runTestProvider'),
+    cancelTestProvider: fn().mockName('api::cancelTestProvider'),
+    updateTestProviderState: fn().mockName('api::updateTestProviderState'),
+    setTestProviderWatchMode: fn().mockName('api::setTestProviderWatchMode'),
+  },
+};
+
 const meta = {
   component: TestingModule,
+  title: 'Sidebar/TestingModule',
   args: {
     testProviders,
     errorCount: 0,
@@ -59,11 +85,11 @@ const meta = {
     warningCount: 0,
     warningsActive: false,
     setWarningsActive: fn(),
-    onRunTests: fn(),
-    onCancelTests: fn(),
-    onSetWatchMode: fn(),
   },
   decorators: [
+    (storyFn) => (
+      <ManagerContext.Provider value={managerContext}>{storyFn()}</ManagerContext.Provider>
+    ),
     (StoryFn) => (
       <div style={{ maxWidth: 232 }}>
         <StoryFn />
@@ -78,10 +104,11 @@ type Story = StoryObj<typeof meta>;
 
 export const Default: Story = {};
 
-export const Collapsed: Story = {
+export const Expanded: Story = {
   play: async ({ canvas }) => {
-    const button = await canvas.findByRole('button', { name: /Collapse/ });
-    await userEvent.click(button);
+    const button = await canvas.findByRole('button', { name: /Expand/ });
+    await fireEvent.click(button);
+    await new Promise((resolve) => setTimeout(resolve, 500));
   },
 };
 
@@ -90,6 +117,7 @@ export const Statuses: Story = {
     errorCount: 14,
     warningCount: 42,
   },
+  play: Expanded.play,
 };
 
 export const ErrorsActive: Story = {
@@ -97,6 +125,7 @@ export const ErrorsActive: Story = {
     ...Statuses.args,
     errorsActive: true,
   },
+  play: Expanded.play,
 };
 
 export const WarningsActive: Story = {
@@ -104,6 +133,7 @@ export const WarningsActive: Story = {
     ...Statuses.args,
     warningsActive: true,
   },
+  play: Expanded.play,
 };
 
 export const BothActive: Story = {
@@ -112,28 +142,29 @@ export const BothActive: Story = {
     errorsActive: true,
     warningsActive: true,
   },
+  play: Expanded.play,
 };
 
 export const CollapsedStatuses: Story = {
   args: Statuses.args,
-  play: Collapsed.play,
 };
 
 export const Running: Story = {
   args: {
     testProviders: [{ ...testProviders[0], running: true }, ...testProviders.slice(1)],
   },
+  play: Expanded.play,
 };
 
 export const RunningAll: Story = {
   args: {
     testProviders: testProviders.map((tp) => ({ ...tp, running: !!tp.runnable })),
   },
+  play: Expanded.play,
 };
 
 export const CollapsedRunning: Story = {
   args: RunningAll.args,
-  play: Collapsed.play,
 };
 
 export const Cancellable: Story = {
@@ -143,6 +174,7 @@ export const Cancellable: Story = {
       ...testProviders.slice(1),
     ],
   },
+  play: Expanded.play,
 };
 
 export const Cancelling: Story = {
@@ -152,12 +184,14 @@ export const Cancelling: Story = {
       ...testProviders.slice(1),
     ],
   },
+  play: Expanded.play,
 };
 
 export const Watching: Story = {
   args: {
     testProviders: [{ ...testProviders[0], watching: true }, ...testProviders.slice(1)],
   },
+  play: Expanded.play,
 };
 
 export const Failing: Story = {
@@ -167,12 +201,14 @@ export const Failing: Story = {
       ...testProviders.slice(1),
     ],
   },
+  play: Expanded.play,
 };
 
 export const Failed: Story = {
   args: {
     testProviders: [{ ...testProviders[0], failed: true }, ...testProviders.slice(1)],
   },
+  play: Expanded.play,
 };
 
 export const Crashed: Story = {
@@ -180,11 +216,31 @@ export const Crashed: Story = {
     testProviders: [
       {
         ...testProviders[0],
-        title: () => "Component tests didn't complete",
-        description: () => 'Problems!',
+        render: () => (
+          <TestProvider>
+            Component tests didn't complete
+            <br />
+            Problems!
+          </TestProvider>
+        ),
         crashed: true,
       },
       ...testProviders.slice(1),
     ],
+  },
+  play: Expanded.play,
+};
+
+export const Updated: Story = {
+  args: {},
+  play: async (context) => {
+    await Expanded.play!(context);
+    triggerUpdate?.();
+  },
+};
+
+export const NoTestProvider: Story = {
+  args: {
+    testProviders: [],
   },
 };