+++ code/frameworks/angular/template/stories/basics/component-with-complex-selectors/attribute-selector.component.ts
@@ -1,6 +1,7 @@
 import { ComponentFactoryResolver, ElementRef, Component } from '@angular/core';
 
 @Component({
+  standalone: false,
   selector: 'storybook-attribute-selector[foo=bar]',
   template: `<h3>Attribute selector</h3>
 Selector: {{ selectors }} <br />